move referendum dropdown to separate file
